+++ src/main/java/com/gugucon/shopping/member/argumentresolver/annotation/MemberId.java
@@ -1,0 +1,12 @@
+package com.gugucon.shopping.member.argumentresolver.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface MemberId {
+
+}
