+++ src/main/java/com/gugucon/shopping/member/controlller/LoginPageController.java
@@ -1,0 +1,15 @@
+package com.gugucon.shopping.member.controlller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/login")
+public class LoginPageController {
+
+    @GetMapping
+    public String loginPage() {
+        return "login";
+    }
+}
