+++ src/main/java/com/gugucon/shopping/member/domain/vo/Nickname.java
@@ -1,0 +1,16 @@
+package com.gugucon.shopping.member.domain.vo;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import lombok.*;
+
+@Embeddable
+@Getter
+@EqualsAndHashCode
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Nickname {
+
+    @Column(name = "nickname")
+    private String value;
+}
