+++ src/main/java/com/gugucon/shopping/member/dto/response/LoginResponse.java
@@ -1,0 +1,21 @@
+package com.gugucon.shopping.member.dto.response;
+
+public final class LoginResponse {
+
+    private String accessToken;
+
+    private LoginResponse() {
+    }
+
+    private LoginResponse(final String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public static LoginResponse from(final String accessToken) {
+        return new LoginResponse(accessToken);
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+}
