--- src/main/java/com/gugucon/shopping/user/argumentresolver/UserIdArgumentResolver.java
@@ -1,29 +1,0 @@
-package com.gugucon.shopping.user.argumentresolver;
-
-import java.util.Objects;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.core.MethodParameter;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-import com.gugucon.shopping.user.argumentresolver.annotation.UserId;
-
-public class UserIdArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private static final String USER_ID = "userId";
-
-    @Override
-    public boolean supportsParameter(final MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(UserId.class) && Objects.equals(
-            parameter.getParameterType(), Long.class);
-    }
-
-    @Override
-    public Long resolveArgument(final MethodParameter parameter,
-        final ModelAndViewContainer mavContainer,
-        final NativeWebRequest webRequest, final WebDataBinderFactory binderFactory) {
-        final HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-        return (Long) request.getAttribute(USER_ID);
-    }
-}
