--- src/main/java/com/gugucon/shopping/user/argumentresolver/annotation/UserId.java
@@ -1,12 +1,0 @@
-package com.gugucon.shopping.user.argumentresolver.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface UserId {
-
-}
