--- src/main/java/com/gugucon/shopping/user/domain/entity/User.java
@@ -1,38 +1,0 @@
-package com.gugucon.shopping.user.domain.entity;
-
-import com.gugucon.shopping.common.domain.entity.BaseTimeEntity;
-import com.gugucon.shopping.user.domain.vo.Email;
-import com.gugucon.shopping.user.domain.vo.Nickname;
-import com.gugucon.shopping.user.domain.vo.Password;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "users")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class User extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Embedded
-    private Email email;
-    @Embedded
-    private Password password;
-    @Embedded
-    private Nickname nickname;
-
-    public User(final Long id, final String email, final String password) {
-        this.id = id;
-        this.email = new Email(email);
-        this.password = new Password(password);
-    }
-
-    public User(final String email, final String password) {
-        this(null, email, password);
-    }
-}
