--- src/main/java/com/gugucon/shopping/user/domain/vo/Nickname.java
@@ -1,16 +1,0 @@
-package com.gugucon.shopping.user.domain.vo;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.*;
-
-@Embeddable
-@Getter
-@EqualsAndHashCode
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Nickname {
-
-    @Column(name = "nickname")
-    private String value;
-}
