--- src/main/java/com/gugucon/shopping/user/domain/vo/Password.java
@@ -1,58 +1,0 @@
-package com.gugucon.shopping.user.domain.vo;
-
-import java.util.Objects;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import com.gugucon.shopping.common.exception.ErrorCode;
-import com.gugucon.shopping.common.exception.ShoppingException;
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-
-@Embeddable
-public class Password {
-
-    private static final String PASSWORD_REGEX = "^(?=.*[a-z]).{4,20}$";
-    private static final Pattern PASSWORD_PATTERN = Pattern.compile(PASSWORD_REGEX);
-
-    @Column(name = "password")
-    private String value;
-
-    protected Password() {
-    }
-
-    public Password(final String value) {
-        validatePattern(value);
-
-        this.value = value;
-    }
-
-    public boolean hasValue(final String value) {
-        return Objects.equals(this.value, value);
-    }
-
-    private static void validatePattern(final String value) {
-        final Matcher matcher = PASSWORD_PATTERN.matcher(value);
-
-        if (!matcher.matches()) {
-            throw new ShoppingException(ErrorCode.INVALID_PASSWORD_PATTERN);
-        }
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        final Password password = (Password) o;
-        return Objects.equals(value, password.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
-    }
-}
