--- src/main/java/com/gugucon/shopping/user/dto/request/LoginRequest.java
@@ -1,23 +1,0 @@
-package com.gugucon.shopping.user.dto.request;
-
-public final class LoginRequest {
-
-    private String email;
-    private String password;
-
-    private LoginRequest() {
-    }
-
-    public LoginRequest(final String email, final String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
