--- src/main/java/com/gugucon/shopping/user/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.gugucon.shopping.user.repository;
-
-import java.util.Optional;
-
-import com.gugucon.shopping.user.domain.vo.Email;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import com.gugucon.shopping.user.domain.entity.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(Email email);
-}
