+++ src/main/resources/data.sql
@@ -5,7 +5,7 @@ values (2, '피자', 'pizza.png', 100, 'test_description', 25000);
 insert into product(id, name, image_file_name, stock, description, price)
 values (3, '사케', 'sake.png', 100, 'test_description', 30000);
 
-insert into users(id, email, password, nickname)
+insert into member(id, email, password, nickname)
 values (1, 'test_email@woowafriends.com', 'test_password!', 'tester1');
-insert into users(id, email, password, nickname)
+insert into member(id, email, password, nickname)
 values (2, 'other_test_email@woowafriends.com', 'test_password!', 'tester2');
