+++ src/test/java/com/gugucon/shopping/integration/CartIntegrationTest.java
@@ -15,8 +15,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import com.gugucon.shopping.TestUtils;
-import com.gugucon.shopping.user.dto.request.LoginRequest;
-import com.gugucon.shopping.user.dto.response.LoginResponse;
+import com.gugucon.shopping.member.dto.request.LoginRequest;
+import com.gugucon.shopping.member.dto.response.LoginResponse;
 import com.gugucon.shopping.item.dto.request.CartItemInsertRequest;
 import com.gugucon.shopping.item.dto.request.CartItemUpdateRequest;
 import com.gugucon.shopping.item.dto.response.CartItemResponse;
