+++ src/test/java/com/gugucon/shopping/integration/OrderIntegrationTest.java
@@ -19,8 +19,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import com.gugucon.shopping.user.dto.request.LoginRequest;
-import com.gugucon.shopping.user.dto.response.LoginResponse;
+import com.gugucon.shopping.member.dto.request.LoginRequest;
+import com.gugucon.shopping.member.dto.response.LoginResponse;
 import com.gugucon.shopping.item.dto.request.CartItemInsertRequest;
 import com.gugucon.shopping.item.dto.response.CartItemResponse;
 import com.gugucon.shopping.order.dto.response.OrderDetailResponse;
