+++ src/test/java/com/gugucon/shopping/item/service/CartServiceTest.java
@@ -8,7 +8,6 @@ import com.gugucon.shopping.item.dto.request.CartItemUpdateRequest;
 import com.gugucon.shopping.item.dto.response.CartItemResponse;
 import com.gugucon.shopping.item.repository.CartItemRepository;
 import com.gugucon.shopping.item.repository.ProductRepository;
-import com.gugucon.shopping.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -41,14 +40,14 @@ class CartServiceTest {
     @DisplayName("장바구니에 상품을 추가한다.")
     void insertCartItem() {
         /* given */
-        final Long userId = 1L;
+        final Long memberId = 1L;
         final Product product = createProduct("치킨", 10000);
         final CartItemInsertRequest cartRequest = new CartItemInsertRequest(product.getId());
         when(productRepository.findById(product.getId())).thenReturn(Optional.of(product));
-        when(cartItemRepository.existsByUserIdAndProductId(userId, product.getId())).thenReturn(false);
+        when(cartItemRepository.existsByMemberIdAndProductId(memberId, product.getId())).thenReturn(false);
 
         /* when */
-        cartService.insertCartItem(cartRequest, userId);
+        cartService.insertCartItem(cartRequest, memberId);
 
         /* then */
         verify(cartItemRepository).save(any());
@@ -58,16 +57,16 @@ class CartServiceTest {
     @DisplayName("장바구니 상품을 조회한다.")
     void readCartItems() {
         /* given */
-        final Long userId = 1L;
+        final Long memberId = 1L;
         final Product chicken = createProduct("치킨", 10000);
         final Product pizza = createProduct("피자", 20000);
-        final CartItem cartItemChicken = new CartItem(1L, userId, chicken, 1);
-        final CartItem cartItemPizza = new CartItem(2L, userId, pizza, 1);
+        final CartItem cartItemChicken = new CartItem(1L, memberId, chicken, 1);
+        final CartItem cartItemPizza = new CartItem(2L, memberId, pizza, 1);
         final List<CartItem> cartItems = List.of(cartItemChicken, cartItemPizza);
-        when(cartItemRepository.findByUserId(userId)).thenReturn(cartItems);
+        when(cartItemRepository.findByMemberId(memberId)).thenReturn(cartItems);
 
         /* when */
-        final List<CartItemResponse> cartItemResponses = cartService.getCartItems(userId);
+        final List<CartItemResponse> cartItemResponses = cartService.getCartItems(memberId);
 
         /* then */
         final List<String> productNames = cartItemResponses.stream()
@@ -82,16 +81,16 @@ class CartServiceTest {
     @DisplayName("장바구니 상품 수량을 수정한다.")
     void updateCartItem() {
         /* given */
-        final Long userId = 1L;
+        final Long memberId = 1L;
         final Product product = createProduct("치킨", 10000);
-        final CartItem cartItem = new CartItem(1L, userId, product, 1);
+        final CartItem cartItem = new CartItem(1L, memberId, product, 1);
 
         final int updateQuantity = 3;
         when(cartItemRepository.findById(cartItem.getId())).thenReturn(Optional.of(cartItem));
 
         /* when */
         cartService.updateCartItemQuantity(cartItem.getId(),
-                new CartItemUpdateRequest(updateQuantity), userId);
+                new CartItemUpdateRequest(updateQuantity), memberId);
 
         /* then */
         assertThat(cartItem.getQuantity()).isEqualTo(new Quantity(updateQuantity));
@@ -101,14 +100,14 @@ class CartServiceTest {
     @DisplayName("장바구니 상품을 삭제한다.")
     void deleteCartItem() {
         /* given */
-        final Long userId = 1L;
+        final Long memberId = 1L;
         final Product product = createProduct("치킨", 10000);
-        final CartItem cartItem = new CartItem(1L, userId, product, 1);
+        final CartItem cartItem = new CartItem(1L, memberId, product, 1);
 
         when(cartItemRepository.findById(cartItem.getId())).thenReturn(Optional.of(cartItem));
 
         /* when */
-        cartService.removeCartItem(cartItem.getId(), userId);
+        cartService.removeCartItem(cartItem.getId(), memberId);
 
         /* then */
         verify(cartItemRepository).delete(cartItem);
