+++ src/test/java/com/gugucon/shopping/order/service/OrderServiceTest.java
@@ -1,7 +1,6 @@
 package com.gugucon.shopping.order.service;
 
-import com.gugucon.shopping.order.service.OrderService;
-import com.gugucon.shopping.user.repository.UserRepository;
+import com.gugucon.shopping.member.repository.MemberRepository;
 import com.gugucon.shopping.item.domain.entity.CartItem;
 import com.gugucon.shopping.order.domain.entity.Order;
 import com.gugucon.shopping.order.domain.vo.ExchangeRate;
@@ -19,7 +18,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.List;
 
 import static com.gugucon.shopping.TestUtils.createProduct;
-import static com.gugucon.shopping.TestUtils.createUser;
+import static com.gugucon.shopping.TestUtils.createMember;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.verify;
@@ -29,7 +28,7 @@ import static org.mockito.Mockito.verify;
 class OrderServiceTest {
 
     @Mock
-    UserRepository userRepository;
+    MemberRepository memberRepository;
     @Mock
     OrderRepository orderRepository;
     @Mock
@@ -45,20 +44,20 @@ class OrderServiceTest {
     @DisplayName("주문에 성공한다.")
     void orderSuccess() {
         /* given */
-        final Long userId = createUser().getId();
-        final CartItem cartItem1 = new CartItem(1L, userId, createProduct("치킨", 10000),
+        final Long memberId = createMember().getId();
+        final CartItem cartItem1 = new CartItem(1L, memberId, createProduct("치킨", 10000),
                 1);
-        final CartItem cartItem2 = new CartItem(2L, userId, createProduct("피자", 20000),
+        final CartItem cartItem2 = new CartItem(2L, memberId, createProduct("피자", 20000),
                 2);
         final List<CartItem> cartItems = List.of(cartItem1, cartItem2);
         final ExchangeRate exchangeRate = new ExchangeRate(1300);
 
-        doReturn(cartItems).when(cartItemRepository).findByUserId(userId);
+        doReturn(cartItems).when(cartItemRepository).findByMemberId(memberId);
         doReturn(exchangeRate).when(exchangeRateProvider).fetchExchangeRate();
-        doReturn(new Order(1L, userId, Order.OrderStatus.ORDERED, exchangeRate)).when(orderRepository).save(any());
+        doReturn(new Order(1L, memberId, Order.OrderStatus.ORDERED, exchangeRate)).when(orderRepository).save(any());
 
         /* when */
-        orderService.order(userId);
+        orderService.order(memberId);
 
         /* then */
         verify(cartItemRepository).deleteAll(cartItems);
